// apps/web/src/lib/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type NodeWithFeaturedImageToMediaItemConnectionEdge = Edge & MediaItemConnectionEdge & OneToOneConnection & {
  __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: MediaItem;
};
// ...
export type Post = ContentNode & DatabaseIdentifier & MenuItemLinkable & Node & NodeWithAuthor & NodeWithComments & NodeWithContentEditor & NodeWithExcerpt & NodeWithFeaturedImage & NodeWithRevisions & NodeWithTemplate & NodeWithTitle & NodeWithTrackbacks & Previewable & UniformResourceIdentifiable & WithAcfArticleFields & WithAcfContentSlices & {
// ...
  categories?: Maybe<PostToCategoryConnection>;
// ...
  tags?: Maybe<PostToTagConnection>;
// ...
};
// ...
export type RootQuery = WithAcfOptionsPageHomepage & WithAcfOptionsPageSiteSettings & {
  __typename?: 'RootQuery';
  /** Entry point to get all settings for the site */
  allSettings?: Maybe<Settings>;
  /** Connection between the RootQuery type and the category type */
  categories?: Maybe<RootQueryToCategoryConnection>;
// ...
  posts?: Maybe<RootQueryToPostConnection>;
// ...
  tags?: Maybe<RootQueryToTagConnection>;
// ...
};
// ...
export type RootQueryToCategoryConnection = CategoryConnection & Connection & {
  __typename?: 'RootQueryToCategoryConnection';
  /** Edges for the RootQueryToCategoryConnection connection */
  edges: Array<RootQueryToCategoryConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Category>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToCategoryConnectionPageInfo;
};
// ...
export type RootQueryToPostConnection = Connection & PostConnection & {
  __typename?: 'RootQueryToPostConnection';
  /** Edges for the RootQueryToPostConnection connection */
  edges: Array<RootQueryToPostConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Post>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToPostConnectionPageInfo;
};
// ...
export type RootQueryToTagConnection = Connection & TagConnection & {
  __typename?: 'RootQueryToTagConnection';
  /** Edges for the RootQueryToTagConnection connection */
  edges: Array<RootQueryToTagConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Tag>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToTagConnectionPageInfo;
};
// ...
export type GetAllCategoriesQueryVariables = Exact<{ [key: string]: never; }>;


export type GetAllCategoriesQuery = { __typename?: 'RootQuery', categories?: { __typename?: 'RootQueryToCategoryConnection', nodes: Array<{ __typename?: 'Category', count?: number | null, databaseId: number, name?: string | null, slug?: string | null }> } | null };

export type GetAllPostsForSitemapQueryVariables = Exact<{
  first: Scalars['Int']['input'];
}>;


export type GetAllPostsForSitemapQuery = { __typename?: 'RootQuery', posts?: { __typename?: 'RootQueryToPostConnection', nodes: Array<{ __typename?: 'Post', slug?: string | null, modified?: string | null }> } | null };

export type GetAllTagsQueryVariables = Exact<{ [key: string]: never; }>;


export type GetAllTagsQuery = { __typename?: 'RootQuery', tags?: { __typename?: 'RootQueryToTagConnection', nodes: Array<{ __typename?: 'Tag', databaseId: number, name?: string | null, slug?: string | null }> } | null };
// ...
export type GetFeaturedPostQuery = { __typename?: 'RootQuery', posts?: { __typename?: 'RootQueryToPostConnection', nodes: Array<{ __typename?: 'Post', databaseId: number, title?: string | null, slug?: string | null, date?: string | null, excerpt?: string | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null, categories?: { __typename?: 'PostToCategoryConnection', nodes: Array<{ __typename?: 'Category', databaseId: number, name?: string | null, slug?: string | null }> } | null, articleFields?: { __typename?: 'ArticleFields', subtitle?: string | null, featured?: boolean | null } | null }> } | null };

export type GetHomepageSettingsQueryVariables = Exact<{ [key: string]: never; }>;


export type GetHomepageSettingsQuery = { __typename?: 'RootQuery', homepage?: { __typename?: 'Homepage', homepageSettings?: { __typename?: 'HomepageSettings', homepageLayout?: Array<string | null> | null, homepageFeaturedOverride?: { __typename?: 'AcfContentNodeConnection', nodes: Array<
          | { __typename?: 'MediaItem' }
          | { __typename?: 'Page' }
          | { __typename?: 'Post', databaseId: number }
        > } | null, homepageHeroSidebarOverrides?: Array<{ __typename?: 'HomepageSettingsHomepageHeroSidebarOverrides', post?: { __typename?: 'AcfContentNodeConnection', nodes: Array<
            | { __typename?: 'MediaItem' }
            | { __typename?: 'Page' }
            | { __typename?: 'Post', databaseId: number }
          > } | null } | null> | null, homepageBannerOverride?: { __typename?: 'AcfContentNodeConnection', nodes: Array<
          | { __typename?: 'MediaItem' }
          | { __typename?: 'Page' }
          | { __typename?: 'Post', databaseId: number }
        > } | null } | null } | null };

export type GetLatestPostsQueryVariables = Exact<{
  first: Scalars['Int']['input'];
  notIn?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>> | InputMaybe<Scalars['ID']['input']>>;
}>;


export type GetLatestPostsQuery = { __typename?: 'RootQuery', posts?: { __typename?: 'RootQueryToPostConnection', nodes: Array<{ __typename?: 'Post', databaseId: number, title?: string | null, slug?: string | null, date?: string | null, excerpt?: string | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null, categories?: { __typename?: 'PostToCategoryConnection', nodes: Array<{ __typename?: 'Category', databaseId: number, name?: string | null, slug?: string | null }> } | null, articleFields?: { __typename?: 'ArticleFields', subtitle?: string | null, featured?: boolean | null } | null }> } | null };
// ...
export type GetPostByIdQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type GetPostByIdQuery = { __typename?: 'RootQuery', post?: { __typename?: 'Post', databaseId: number, title?: string | null, slug?: string | null, date?: string | null, excerpt?: string | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null, categories?: { __typename?: 'PostToCategoryConnection', nodes: Array<{ __typename?: 'Category', databaseId: number, name?: string | null, slug?: string | null }> } | null, articleFields?: { __typename?: 'ArticleFields', subtitle?: string | null, featured?: boolean | null } | null } | null };
// ...
export type GetPostsByCategoryIdQueryVariables = Exact<{
  first: Scalars['Int']['input'];
  categoryId?: InputMaybe<Scalars['Int']['input']>;
  notIn?: InputMaybe<Array<InputMaybe<Scalars['ID']['input']>> | InputMaybe<Scalars['ID']['input']>>;
}>;


export type GetPostsByCategoryIdQuery = { __typename?: 'RootQuery', posts?: { __typename?: 'RootQueryToPostConnection', nodes: Array<{ __typename?: 'Post', databaseId: number, title?: string | null, slug?: string | null, date?: string | null, excerpt?: string | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null, categories?: { __typename?: 'PostToCategoryConnection', nodes: Array<{ __typename?: 'Category', databaseId: number, name?: string | null, slug?: string | null }> } | null, articleFields?: { __typename?: 'ArticleFields', subtitle?: string | null, featured?: boolean | null } | null }> } | null };
// ...
export const GetAllCategoriesDocument = new TypedDocumentString(`
    query GetAllCategories {
  categories(where: {hideEmpty: true}, first: 20) {
    nodes {
      ...CategoryFields
      count
    }
  }
}
    fragment CategoryFields on Category {
  databaseId
  name
  slug
}`) as unknown as TypedDocumentString<GetAllCategoriesQuery, GetAllCategoriesQueryVariables>;
export const GetAllPostsForSitemapDocument = new TypedDocumentString(`
    query GetAllPostsForSitemap($first: Int!) {
  posts(first: $first, where: {status: PUBLISH}) {
    nodes {
      slug
      modified
    }
  }
}
    `) as unknown as TypedDocumentString<GetAllPostsForSitemapQuery, GetAllPostsForSitemapQueryVariables>;
export const GetAllTagsDocument = new TypedDocumentString(`
    query GetAllTags {
  tags(where: {hideEmpty: true, orderby: COUNT, order: DESC}, first: 15) {
    nodes {
      ...TagFields
    }
  }
}
    fragment TagFields on Tag {
  databaseId
  name
  slug
}`) as unknown as TypedDocumentString<GetAllTagsQuery, GetAllTagsQueryVariables>;
// ...
export const GetFeaturedPostDocument = new TypedDocumentString(`
    query GetFeaturedPost {
  posts(where: {orderby: {field: DATE, order: DESC}}, first: 1) {
    nodes {
      ...PostCardFields
    }
  }
}
    fragment CategoryFields on Category {
  databaseId
  name
  slug
}
fragment MediaItemFields on MediaItem {
  databaseId
  sourceUrl
  altText
  mediaDetails {
    width
    height
  }
}
fragment PostCardFields on Post {
  databaseId
  title
  slug
  date
  excerpt
  featuredImage {
    node {
      ...MediaItemFields
    }
  }
  categories {
    nodes {
      ...CategoryFields
    }
  }
  articleFields {
    subtitle
    featured
  }
}`) as unknown as TypedDocumentString<GetFeaturedPostQuery, GetFeaturedPostQueryVariables>;
export const GetHomepageSettingsDocument = new TypedDocumentString(`
    query GetHomepageSettings {
  homepage {
    homepageSettings {
      homepageFeaturedOverride {
        nodes {
          ... on Post {
            databaseId
          }
        }
      }
      homepageLayout
      homepageHeroSidebarOverrides {
        post {
          nodes {
            ... on Post {
              databaseId
            }
          }
        }
      }
      homepageBannerOverride {
        nodes {
          ... on Post {
            databaseId
          }
        }
      }
    }
  }
}
    `) as unknown as TypedDocumentString<GetHomepageSettingsQuery, GetHomepageSettingsQueryVariables>;
export const GetLatestPostsDocument = new TypedDocumentString(`
    query GetLatestPosts($first: Int!, $notIn: [ID]) {
  posts(
    where: {orderby: {field: DATE, order: DESC}, notIn: $notIn}
    first: $first
  ) {
    nodes {
      ...PostCardFields
    }
  }
}
    fragment CategoryFields on Category {
  databaseId
  name
  slug
}
fragment MediaItemFields on MediaItem {
  databaseId
  sourceUrl
  altText
  mediaDetails {
    width
    height
  }
}
fragment PostCardFields on Post {
  databaseId
  title
  slug
  date
  excerpt
  featuredImage {
    node {
      ...MediaItemFields
    }
  }
  categories {
    nodes {
      ...CategoryFields
    }
  }
  articleFields {
    subtitle
    featured
  }
}`) as unknown as TypedDocumentString<GetLatestPostsQuery, GetLatestPostsQueryVariables>;
// ...
export const GetPostByIdDocument = new TypedDocumentString(`
    query GetPostById($id: ID!) {
  post(id: $id, idType: DATABASE_ID) {
    ...PostCardFields
  }
}
    fragment CategoryFields on Category {
  databaseId
  name
  slug
}
fragment MediaItemFields on MediaItem {
  databaseId
  sourceUrl
  altText
  mediaDetails {
    width
    height
  }
}
fragment PostCardFields on Post {
  databaseId
  title
  slug
  date
  excerpt
  featuredImage {
    node {
      ...MediaItemFields
    }
  }
  categories {
    nodes {
      ...CategoryFields
    }
  }
  articleFields {
    subtitle
    featured
  }
}`) as unknown as TypedDocumentString<GetPostByIdQuery, GetPostByIdQueryVariables>;
// ...
export const GetPostsByCategoryIdDocument = new TypedDocumentString(`
    query GetPostsByCategoryId($first: Int!, $categoryId: Int, $notIn: [ID]) {
  posts(
    where: {categoryId: $categoryId, orderby: {field: DATE, order: DESC}, notIn: $notIn}
    first: $first
  ) {
    nodes {
      ...PostCardFields
    }
  }
}
    fragment CategoryFields on Category {
  databaseId
  name
  slug
}
fragment MediaItemFields on MediaItem {
  databaseId
  sourceUrl
  altText
  mediaDetails {
    width
    height
  }
}
fragment PostCardFields on Post {
  databaseId
  title
  slug
  date
  excerpt
  featuredImage {
    node {
      ...MediaItemFields
    }
  }
  categories {
    nodes {
      ...CategoryFields
    }
  }
  articleFields {
    subtitle
    featured
  }
}`) as unknown as TypedDocumentString<GetPostsByCategoryIdQuery, GetPostsByCategoryIdQueryVariables>;